Completed .4
Completed the document viewing system, players can now view multiple documents with different texts. Players aren't able to move when viewing a document, they are given clear guidance in terms of button prompts, so they know how to exit or inspect the document to view the documents text in a more legible font.